+++ pkg/query/logical/measure/measure_analyzer.go
@@ -93,6 +93,23 @@ func Analyze(criteria *measurev1.QueryRequest, metadata []*commonv1.Metadata, ss
 		}
 	}
 
+	ms := s.(*schema)
+	if len(tagProjection) > 0 {
+		if err := ms.common.ValidateProjectionTags(tagProjection...); err != nil {
+			return nil, err
+		}
+	}
+	fieldProjection := criteria.GetFieldProjection().GetNames()
+	if len(fieldProjection) > 0 {
+		projFields := make([]*logical.Field, len(fieldProjection))
+		for i, fieldName := range fieldProjection {
+			projFields[i] = logical.NewField(fieldName)
+		}
+		if err := ms.ValidateProjectionFields(projFields...); err != nil {
+			return nil, err
+		}
+	}
+
 	// parse limit and offset
 	limitParameter := criteria.GetLimit()
 	if limitParameter == 0 {