+++ CHANGES.md
@@ -8,7 +8,8 @@ Release Notes.
 
 - Remove Bloom filter for dictionary-encoded tags.
 - Implement BanyanDB MCP.
-- Remove check requiring tags in criteria to be present in projection
+- Support deleting non-entity tags when updating the schema.
+- Remove check requiring tags in criteria to be present in projection.
 - Add sorted query support for the Property.
 - Update bydbQL to add sorted query support for the Property.
 - Remove the windows arch for binary and docker image.